Rename files and exported functions
