+++ public/app/plugins/datasource/loki/metricTimeSplitting.test.ts
@@ -1,4 +1,4 @@
-import { getRangeChunks } from './metricTimeChunking';
+import { getRangePartition } from './metricTimeSplitting';
 
 describe('metric getRangeChunks', () => {
   it('should split time range into chunks', () => {
@@ -6,7 +6,7 @@ describe('metric getRangeChunks', () => {
     const end = Date.parse('2022-02-06T14:11:03');
     const step = 10 * 1000;
 
-    expect(getRangeChunks(start, end, step, 25000)).toStrictEqual([
+    expect(getRangePartition(start, end, step, 25000)).toStrictEqual([
       [Date.parse('2022-02-06T14:10:00'), Date.parse('2022-02-06T14:10:10')],
       [Date.parse('2022-02-06T14:10:20'), Date.parse('2022-02-06T14:10:40')],
       [Date.parse('2022-02-06T14:10:50'), Date.parse('2022-02-06T14:11:10')],
@@ -17,6 +17,6 @@ describe('metric getRangeChunks', () => {
     const start = Date.parse('2022-02-06T14:10:03');
     const end = Date.parse('2022-02-06T14:10:33');
     const step = 10 * 1000;
-    expect(getRangeChunks(start, end, step, 1000)).toEqual([[start, end]]);
+    expect(getRangePartition(start, end, step, 1000)).toEqual([[start, end]]);
   });
 });
